agents/posthog-digest/requirements.txt
aiohttp>=3.9.0
discord.py>=2.3.0
tenacity>=8.2.0
